tomogachi/backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-generativeai==0.3.2
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.26.4
langchain==0.1.0
langchain-google-genai==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23